+++ tests/utils.py
@@ -62,7 +62,7 @@ from crate.operator.utils.kubeapi import (
 
 logger = logging.getLogger(__name__)
 
-CRATE_VERSION = "5.0.0"
+CRATE_VERSION = "5.2.5"
 DEFAULT_TIMEOUT = 60